+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release Version 0.17
 
 **Updates**
-* Changed the ComboBox of the swing-EditorSubClass to a type that can filter.
+* Changed the ComboBox of EditorSubClass to a type that can filter.
 * Added option to EditorSubClass to not shorten class names.